Move null check to _get_secrets
The check for "None" is better placed inside the get_secrets function,
with this, get_secrets will always return a list, either empty list, or
the list with the secrets from the yaml. Changed checks accordingly.
Changed default get to be an empty list if the "secret" key is not
present in the secret values file.