allow var hide_sensitive_output to disable no_log in debug environments
Adds the var hide_sensitive_output eveywhere we have no_log so that sensitive data like
secrets are hidden by default. Setting this var to `false` allows us temporarily disable
the no_log protection when we are debugging in secure environments.

An example of how we can use this is to set
```sh
export EXTRA_PLAYBOOK_OPTS='-e hide_sensitive_output="false" -vvv'
```
before we run a command like `./pattern.sh make load-secrets`.
(Do note, however, that we currently do not actually propagate this environment variable
to the `ansible-playbook` command. A separate PR is required to update our
common module to propagate this value to all of the playbooks.)