Reduce the usage of the acm secret filter and support ACM 2.13
ACM 2.13 uses managed service accounts so there is no <spoke>-cluster
secret anymore to parse. Instead we look for a secret called
'application-manager' inside the namespace for the spoke cluster.

That secret contains the bearerToken. We still keep 2.12 compatibility
and only run this when we cannot find the <spoke>-cluster secret.

Tested on both ACM 2.13 and 2.12